+++ eventloop.go
@@ -230,7 +230,7 @@ func handleMessageFromPubSub(meta *pubSubMetadata, channel, payload string) erro
 	}
 
 	for _, sub := range subs {
-		invokeCallback(meta.lifecycleCtx, meta.logger, meta.recorder, meta.workerPool, meta.callbackWg, sub.callback, msg)
+		invokeCallbackOrdered(meta.lifecycleCtx, meta.logger, meta.recorder, meta.workerPool, meta.callbackWg, sub, msg)
 	}
 	return nil
 }
@@ -257,7 +257,7 @@ func handlePMessageFromPubSub(meta *pubSubMetadata, pattern, channel, payload st
 	}
 
 	for _, sub := range subs {
-		invokeCallback(meta.lifecycleCtx, meta.logger, meta.recorder, meta.workerPool, meta.callbackWg, sub.callback, msg)
+		invokeCallbackOrdered(meta.lifecycleCtx, meta.logger, meta.recorder, meta.workerPool, meta.callbackWg, sub, msg)
 	}
 	return nil
 }
@@ -286,7 +286,7 @@ func handleSMessageFromPubSub(meta *pubSubMetadata, channel, payload string) err
 	}
 
 	for _, sub := range subs {
-		invokeCallback(meta.lifecycleCtx, meta.logger, meta.recorder, meta.workerPool, meta.callbackWg, sub.callback, msg)
+		invokeCallbackOrdered(meta.lifecycleCtx, meta.logger, meta.recorder, meta.workerPool, meta.callbackWg, sub, msg)
 	}
 	return nil
 }
@@ -334,6 +334,6 @@ func notifySubscriptionsOfFailure(meta *pubSubMetadata, err error) {
 			Timestamp:        time.Now(),
 			SubscriptionType: sub.subType,
 		}
-		invokeCallback(meta.lifecycleCtx, meta.logger, meta.recorder, meta.workerPool, meta.callbackWg, sub.callback, msg)
+		invokeCallbackOrdered(meta.lifecycleCtx, meta.logger, meta.recorder, meta.workerPool, meta.callbackWg, sub, msg)
 	}
 }